loader: actually set serial flag when using serial console
In F15 and earlier, if the primary console was a serial port, loader
would set the appropriate flag (LOADER_FLAGS_SERIAL). This got dropped
accidentally in the refactoring work after F15 (see commit 7605892).

This will correctly set the flag (and thus force text mode) if
init_serial() indicates that a serial console is being used.