Start storage probing in its own separate thread.
In this way, we have all the probing done in the background while the UI is
instantiated and displayed.  It's the intention that this process is complete
by the time the network spoke is displayed, though I currently have no way
to enforce that.

This code is likely to move and become significantly fancier in the future,
but it works in testing for now.

Additionally, a storage change is required.  We need to ensure that there is
only one DeviceTree object ever created and set as storage.devicetree.  If
not, the instance referenced in the UI code will not be the same as the one
that's been populated.  So, add a reset method that gets called from
storage.reset that does all the same stuff but keeps us in the same instance.