Update ppc SMS bios after installation
Here is a proposed patch that updates a PowerPC's BIOS to boot from the
newly installed hard disk rather than back to the installer.

The BIOS maintains a list of bootable targets which can be queried with
the nvram utility.  This list can be modified so that the installation
harddrive will show up first.  This code will read the list, figure out
what drive contains the PReP partition, prepends that drive to the list,
and then updates BIOS.

For example the list could be the DVD, hard drive, and ethernet.

[anaconda root@(none) /]# nvram --print-config=boot-device
/vdevice/v-scsi@30000005/disk@8100000000000000
/vdevice/v-scsi@30000002/disk@8100000000000000
/vdevice/l-lan@30000003:speed=auto,duplex=auto,000.000.000.000,[etc.]

After running this function, the list will be as follows:

[anaconda root@(none) /]# nvram --print-config=boot-device
/vdevice/v-scsi@30000002/disk@8100000000000000
/vdevice/v-scsi@30000005/disk@8100000000000000
/vdevice/l-lan@30000003:speed=auto,duplex=auto,000.000.000.000,[etc.]