Don't exit the spoke's main loop from inside the Hub.
The spoke is responsible for doing its own exiting.  Without this patch,
we end up in one of two situations:

(1) If the spoke does a main_quit(), then the main_quit in the Hub will
cause anaconda to exit as soon as the spoke is done.

(2) If the spoke doesn't do a main_quit(), we never really leave the spoke
so its apply method will not get run.  We get the illusion of leaving the
spoke, however, due to it being hidden and the actions on the Hub still
taking effect.