Initialize threads in gobject before starting up storage probing.
If this isn't done, we'll end up in a situation where gobject/gtk/etc. don't
think threads are enabled, so the storage probing thread will only get to
run before the GTK main loop is started, or when you click buttons or
otherwise wake up the main loop.  In other words, it won't really run in the
background.