Update existing UI files for the new base widget layout.
This mainly involved getting rid of alignments and boxes since BaseWindow
will take care of that now.  The only interesting part is in software.ui
where I had to add a box as the only child of the action_area.  This is
because BaseWindow enforces action_area to be vertical but we want the
software spoke to be divided up horizontally.