Clarify how to use builderObjects.
The reason we need this attribute at all is that when we go to load the glade
file, all objects in the file get loaded.  Notably, this includes dialogs
that are their own object hierarchy.  Then when connect_signals() is called,
all signals in all loaded objects will attempt to be hooked up.  However,
because we implement dialogs as their own Python objects, the "self" passed to
connect_signals will not include implementations of those signals.  Thus,
tracebacks.

I could fix this by moving dialogs to their own files, but that doesn't seem
to be more clean.