Add a short class to help manage threads.
It'd be nice to be able to grab a reference to the running anaconda threads
for various reasons, so wrap that up into a little class to make it a bit
more convenient.  Also, I can see that we might want to do more fancy things
with threads in the future so now we have a place to put that kind of code.