fix serial console option parsing (#767745)
Anaconda was expecting serial console options with only the baud rate
after the comma, for example:
  console=ttyS0,115200

But this is also valid (see Documentation/kernel-parameters.txt):
  console=ttyS0,115200n8

So this patch adds a function to parse the options correctly, and improves
the GRUB serial_command() method to use the parsed options wherever they
vary from the default values.

Based on a patch by Jan Stancek <jstancek@redhat.com> - thanks!