Add support to some widgets for retranslating the UI on language change.
GTK doesn't really make this easy on us.  The only thing we can do is tear
everything down and rebuild, which is not at all an optimal solution.
Alternatively, we can maintain lists of original strings and the widgets that
contain them, and do it on demand by hand.

Not everything is supported yet.  I'll add support as needed.