Allow selecting DiskOverviews via the keyboard.
Note the "selected" -> "chosen" change.  "selected" means something specific
to GTK, and it doesn't mean what I thought.  Thus, I am now using "chosen" to
mean you have hit enter/space on the DiskOverview and made it part of the
install.

This also gets rid of the rounded corner piece, since I was unable to make that
work with the keyboard focus (for now).