Support not-yet-ready spokes in the _createBox loop for hubs.
NormalSpokes (those that you can select off a Hub) can be hidden.  Now, they
can also be ready or not.  A Spoke is ready if it has all the information
needed to be displayed.  Most will always be ready.  However, the storage spoke
needs to wait for storageInitialize to finish which could potentially take a
very long time.

However, we do not want to sit in _createBox for a long time as that delays
showing the first hub thus making the UI look hung.  Therefore, we will just
mark not-yet-ready spokes as insensitive until they're ready.