Call our custom exception handler when an exn occurs in a thread.
The python threading module has some goofy custom exception handling behavior
that basically means we'll never see an exception that occurs in a thread.
This is apparently a known bug (see comment in the commit) with a known
workaround, so use it.

Of course, exception handling overall doesn't really work right now.