Very lame initial exception reporting support for the newui branch.
I can't really use the UI parts of python-meh without updating it to GTK3,
which I can't do until firstboot and s-c-ks are also updated.  Currently
attempting to handle an exception results in another exception, which is
not at all useful.  Thus, this patch makes sure the /tmp/anaconda-tb-*
file is written and then prints the exception to tty1 and exits.

Hey, at least we know something happened now.