Add a check method to NormalSpokes, and have it called from on_back_clicked.
Certain spokes (storage somewhat and software selection especially) require that
some basic sanity checks be performed before continuing.  In the case of
software selection, this process could take a very long time.  Also, it needs
to be done before going to the second hub since the user may have to make
changes and we don't allow going back.

The solution, then, is a check method that's run out of yet another thread.
It's started as soon as the Back button is clicked on a spoke.  If the user
goes back into the spoke, we request the thread exit.  If the user clicks
Continue on the hub before checking is done, the hub will need to wait.

A couple things need to be done before this all works:  First, Hubs need a
way of noticing that check threads are still running and wait for them (with
a notification dialog).  Second, check threads need to be derived from a
class that enforces them having a main loop and kill switch.