Remove the anaconda attribute from Storage and add a bootloader attr.
The main thing we used the anaconda attribute for was access to the
installer's user interface. The approach taken in this patch is to
replace direct usage of a user interface library with optional error
callbacks. The callbacks take an Exception instance and, in some
cases, a StorageDevice instance as arguments. They can display pop-
up windows or whatever they want to do. A return value of True
indicates that the error was fatal. In this case the exception is
raised.