Remove the storage attribute from BootLoader.
This also removes a lot of the "magic" from BootLoader. The drive
list must now be set explicitly using the set_drive_list method.
Similarly, the stage1 device is only set by a call to the
set_stage1_device method when the storage configuration is settled.
Additionally, the default image/target/stanza is no longer generated
on demand if it does not exist when access is requested.