move vncpassword handling into anaconda; remove recoverVNCPassword
loader was writing the vncpassword out to a file, because:

  a) many years ago, VNC got started by loader, and
  b) VNC used to need to read its password from a file.

But now it is the Glorious Future and anaconda can just read the
'vncpassword' boot/cmdline arg and hand it to the VncServer object.

This also removes recoverVNCPassword and pw_init_file from the VncServer
class, and the tests associated with them.