Add support for UEFI Mac installs
Macs are slightly special in terms of their UEFI implementation. The
easiest way to handle boot choice is via the firmware boot picker menu,
which doesn't make use of EFI boot variables. In order for something to
show up here, the filesystem needs to be HFS+ and have some magic done
to it. Handle this case by detecting that we're on a Mac and using an HFS+
/boot/efi partition rather than a FAT one. That lets us then install the
mactel-boot package which will deal with configuring the bootloader.