Set default lang and create default locale files early
If you import locale before the file(s) get created, you end up with a
traceback from locale that says:

  Error: unsupported locale setting

And this happens no matter what the locale is.

We used to create the locale files for the default lang in loader, so
now let's have anaconda declare a default locale and build its locale
data before we import locale.