Add anaconda dracut module [WIP!]
This adds the dracut directory and anaconda-dracut subpackage.

The 'anaconda' dracut module replaces loader. It should handle all the
tasks that loader used to handle in the first stage of the boot process.

It's a work in progress, but it currently supports repo={http,ftp,nfs}
and most anacaconda network options.

Kickstart parsing is handled using a python interpreter and pykickstart,
which (surprisingly) only adds ~600k to the initramfs.

Driver disks are not yet supported.