Add screenshot support back in.
This is just like it was in the old UI, except there's no dialogs telling
you if it worked or not.  Also, you can take a screenshot of anything that
derives from UIObject.  This includes most (but not all) dialogs.

Note that the part that copies screenshots over to the finished system is
still missing, but that fits in with a whole set of post-installation tasks
that we don't have figured out yet.