Add a selector attribute to the related spoke.
Passing in the readyCB means multiple callbacks end up referencing the same
selector.  I could probably do something very clever here, but just adding
the reference is pretty clean.  I'm not sure why I was so opposed to doing
this in the first place.

Note that this also changes the skipTo handling to match.