Support implementing portions of anaconda's post-install as %post scripts.
For now, all that's supported is copying screenshots.  However this could
be extended to other portions that really don't benefit from being in
python (restorecon, etc.).  Of course none of these scripts get run yet
given that we don't have any way to run the kickstart file, but once we
do we'll get this for free.