Pass ifname to dracut for disconnected fcoe ifaces (#743784)
Resolves: rhbz#743784

The case:
We discover LUN on an iface which is part of multipath setup.
If the iface is disconnected after discovery, anaconda doesn't
write dracut ifname argument for the disconnected iface path
(in Network.dracutSetupArgs).
Dracut needs the explicit ifname= because biosdevname
fails to rename the iface (because of BFS booting from it).