Add support for running media check for both ISO images and CD/DVDs.
This requires some weird glib IO channel code to handle updating the progress
bar while a process runs, but it's really not all that bad.  Notifying the user
whether their media is good or not is still rough, but at least there's
something there while we figure out what really should be going on.

I have temporarily removed the Cancel button as that makes things a lot more
difficult and we can always work on it later if we really need it.