Set the default packaging payload as an attribute on Anaconda.
I've first removed the definitions of getBackend from our standard install
classes.  The base class returns None.  For other products with their own
install classes, they can have getBackend return something.  For our
products, we will fall back to looking up the payload class based on flags.

Note that payload.setup() still needs to be called, but this should be done
intentionally when it's needed since it must happen after the network is up
and storage is populated.

Note also that this adds more crud to Anaconda, but we're going to be removing
most of this object later so hopefully it won't be so bad.  We'll also need
to better document install classes.