intelligently choose the window size (#800609)
Picking a maximum window size can be complicated by multiple monitor
setups as well as not knowing which display anaconda will end up on.
This patch solves the window size problem by examining the xrandr output
and picking the smallest width and height available and setting the size
to that. These are picked independently so that the size chosen can fit
on any display.

This means that on a single display it will use all of the available
space. On a multiple monitor setup it will pick a size that can be used
on any of the connected displays.