Populate the software selection spoke with data from the new packaging API.
This isn't the final data we want displayed.  Right now I'm just dumping
desktop groups in the left list and everything else in the right list.  Later
we will want spin-like things in the left list and groups related to those
spin-like things in the right list.

However, this is good enough for now to show that the packaging code works.
Also, the selections made on this spoke will now be propagated into the ksdata
and show back up next time the spoke is visited.