add fedora-import-state.service (fix NFS root: #799989)
dracut writes out an ifcfg-ethX file and a dhclient lease file for the
interface used to mount the runtime. We need to pass that along to
NetworkManager so it won't bring down the interface we're using for our
root device.

This service copies the contents of the /run/initramfs/state directory
into its intended place, as would normally be done in fedora-readonly
for NFS readonly-root systems.