fix inst.ks handling in anaconda
If inst.ks is on the commandline, the dracut module fetches and
preprocesses the kickstart, then saves it to /run/install/ks.cfg.

So if we see a 'ks' option on the boot commandline, anaconda's option parser
should assume that the ksfile is /run/install/ks.cfg.

On the other hand, the '--kickstart' option can still be used on the
commandline to set ksfile to a different filename.