set rd.{luks,dm,md,lvm}=0 unless the user says otherwise
Currently anaconda will crash if we don't set rd.luks=0. The others
anaconda can *mostly* handle but historically they would have been set
to 0, so it's more reliable if we do it that way.

Someday soon we'll make sure anaconda can handle already-active devices
and we'll remove this block of code. Really.