Bring back the initialize callback argument.
Turns out, I need this callback to handle updating the hub's completeness
tracking.  When spokes are first created, the hub checks to see if they
are complete and what their status is for making the grid.  Then for spokes
that take a long time setting up, they fire off a thread and immediately
return.  When that thread finishes, the completeness and status may have
changed but there's no way to update the hub.

Thus, this argument has to return.