fix run_kickstart for the non-repo case
If you're booting from a CD but using a kickstart for other stuff, we
still need to set up / retrigger block stuff *even though you don't have
a repo defined by the kickstart*.

The problem was two things: first, we were leaving $root set to
"anaconda-kickstart", so parse-anaconda-repo.sh didn't change anything.

Second, when root was set to "anaconda-auto-cd" because there wasn't a
inst.repo arg, we never triggered the disk magic. So we never found the
CD.