Format PReP partition
This fixes two bugs.  The first is that the PReP partition is not created
during a text mode install.  The second is that the PReP partition needs
to be zeroed out before grub is run.  Grub verifies that no data exists
on the partition before installation.
(cherry picked from commit 00fa7bcc9e09a96d51039d112c3feadee9ade75a)