dracut/parse-kickstart: handle network --device=link (or none)
If you have a kickstart where your first network line has no device (or
--device=link), we won't have a valid net.device ("" or "link",
respectively.) Still a valid kickstart though, so we need to generate
correct dracut args for that case.

Also we need to set rd.neednet so the device always goes online in
dracut (matching loader's behavior).

This means we don't need to write ifcfg if net.device is empty:
1) empty 'net.device' is only valid for the first device,
2) first device always comes online,
3) online devices get ifcfg written by dracut.

(It also means we don't need to touch /tmp/net.ifaces, since that happens
in step 3.)