Add a first run at the custom partitioning spoke.
This spoke is indirect, meaning you can only get to it through the initial
storage configuration spoke.  It doesn't appear anywhere on the hub.

It doesn't yet do very much useful.  I should be grabbing the devices to display
in the right hand side instead of the list of partitioning requests, and you
can't do any editing anyway.  But, it's something and allows for checking out
how everything looks.