Start populating the UI based on existing devices, not requests.
In the custom partitioning UI, we want to present the user with everything as
it is now so they can then change it to what they want it to be.  Thus, we need
to not look at the requests.

Note that this currently only works for things like swap or biosboot, not real
filesystems as storage.mountpoints is not yet set up.  That's another pile of
work.