Don't write out our internal %post scripts to /root/anaconda-ks.cfg.
Since some parts of anaconda are now implemented as kickstart scripts and
are written right into the ksdata, this means they will also be written to
the kickstart file we output at the end of installation.  Of course, we
don't want to do this as it'll include tons of information the user doesn't
care about and never selected.  So don't include them.