Modify the autopart and bootloader commands to work with the new UI.
First, no more anaconda instances getting passed in to everything.  That's just
sloppy.  Second, the execute methods should be taking action now, not just
setting other data.  We're also not going to be able to call all execute methods
in a big loop because there's an order implied (partitioning before packaging,
at the least) so we need to get rid of that code too.