Split the progress updating method into two.
We need to be able to update the progress bar and the message independently.
The message should be updated when an action begins so you know what's
taking so long, and the progress bar should be updated when an action ends
so you know how far you are through now.