mount before resizing live image (#811706)
In some circumstances the date on the live image could be in the future.
eg. if the system clock was set to last year. e2fsck doesn't like the
future so it would stop the install.

This fixes it by mount/umounting the root fs after it is copied and
before e2fsck and resize are run. The mount sets the m_time to the
current time, preventing e2fsck from losing its mind.