Redo the progress bar code so it actually works.
With the previous code, I could never get the UI to be redrawn to reflect the
changes, even though the changes appeared to have taken effect.

This new code instead uses a message queue for communication between the thread
doing all the hard work and the main thread that handles GTK stuff.  The storage
code will shove messages into that queue when it does something like create a
filesystem, and the main thread watches in the idle loop for messages.

I've also provided yet another context manager to make all this code simpler.