Use a yum cache dir outside the install root.
In the new UI users configure software selection and repos before
the new root filesystem exists on ROOT_PATH. This means we need to
do the configuration with a yum installroot outside of ROOT_PATH.
Of course, we need to switch to the new/actual root before we
install packages.

I am planning to smooth this out such that it will not require a
full run through updateBaseRepo or a storage instance.