Generate an installer yum config for package install.
We have to switch yum over from the configuration-time installroot to
the install-time installroot. This, of course, is not as simple as
changing an attribute of the YumBase instance. Instead, it requires
that we create a new instance of YumBase with the new configuration.
Instead of calling updateBaseRepo again to start from scratch and try
to establish an install-time repo set like we do during configuration,
write out the configuration we want for install time based on what
was configured and then tell the install-time YumBase to use it.