Fix up automatic moving behavior for kickstart installs.
The basic idea is that we skip any standalone spokes that are already filled out
by the input kickstart file, and that we only display the summary hub so that
something's being displayed while background processes work.

Any important spokes not filled out by the input kickstart file will get the
caution icon and the user will have to complete them before continuing.  The
message at the bottom of the summary hub explains that.

This behavior leaves open the possibility that the user will be able to visit a
spoke while background processes work.  In that case, we want to wait until the
user leaves that spoke before automatically continuing.  The alternative is to
move as soon as anaconda is ready, which is really quite jarring.