Fix twisty logic around forcing text mode (#811868)
The existing logic forces text-mode if /usr/bin/Xorg is missing. Except
it doesn't check if you're using VNC first, which caused problems on
certain POWER systems where we're not including Xorg anymore.

We could just add 'and not flags.usevnc' to the if statement but it's
already really, *really* hard to understand what's going on there.

So: this patch cleans up the logic in a way that should be equivalent,
*and* fixes the bug.

A couple notes, if you're checking my work:

- The flags.livecdInstall check was to avoid starting X during liveinst
  (assuming it'd already be running on a live system), so the intent
  there was:
  flags.livecdInstall == flags.preexisting_x11

- The 'isS390' check in the first block was added in commit 4d42471, way
  back in 2004: "enable GUI/VNC installs on mainframe again, there's no
  Xorg binary". So the intent there was:
  isS390 == (flags.preexisting_x11 or flags.usevnc)

Knowing that, you can replace the conditions and do a little juggling
and come up with what I have here. Or something like it.