dracut: fix "repo=cdrom" (also kickstart with "cdrom")
The replacement hack for splitsep wasn't detecting the case where the
repo is just "cdrom" - no ":PATH" at the end - and trying to remove the
"cdrom:" from the front of "cdrom", which doesn't work.

In that case, we'll end up with $rest = $repo, so check for that, and
set rest="".