The welcome spoke needs to remember original strings, too.
Otherwise when the user changes language, they're going to see a string like
"WELCOME TO %s %s" which is of course not right.  I do not yet know whether
more places are going to need this modification but if so that will be the
time to make this code more generic.

Also, it's still hard to test without real translations.