Fix some broken tests by unaliasing modules
On an up-to-date as of an hour ago f17 system, on the newui branch,
with no modifications, running "make unittest" gives me 239 errors
and 14 failures.

The following small patches reduce that to 84 errors and 4 failures,
resolving 155 errors and 10 failures.

Beyond this, the rest of the failures look real, most likely due
to code changes without corresponding test suite changes.