Set timezone in the initialization method, but don't start timers
We need to set the timezone in the initialization method because
the Datetime spoke's status is queried before the refresh method
gets invoked. However the change of timezone results in a change
of the displayed date/time and we don't want to start timers for
updating the date/time when the spoke is not shown.