Network spoke: wait for config objects after device activation
ACTIVATED state of device doesn't ensure immediate availability
of its configuration (NMIP4Config object) so we might need to
make more tries.

Using separate callbacks for updating just config UI connected to
config property change event would be overcomplicated because we are
waiting for two properties (ipv4 and ipv6), and modifications of UI
of one protocol depends on values of both of them.