The software spoke's readiness also depends upon its own watcher thread.
When a spoke is created, initialize should always be called before readiness is
checked.  However for those spokes that spawn off a thread in the initialize
method, readiness could be checked before that thread is done.  We need to make
sure to check that the watcher is still running as well.