Added the basis of the left hand side of the custom partitioning UI.
This removes the previous store and tree view code, adding instead the Accordion
and Page code.  It also adds the Page that prompts you for how to create a new
partitioning layout and makes that the default for now.

So, no partitioning is yet displayed here and the creation button doesn't do
anything.  But it's a start.