Keep a _current_selector attr on the spoke, don't worry with the accordion.
This means the accordion and page methods to get the current selector are
temporarily unused, but I can imagine such code will find a use again.  Thus
I will leave it in for now.

Also, save the format's label as entered from the UI as a test to make sure
this all works.  It still won't make it onto the disk yet.