Giving a MountpointSelector focus should cause it to be displayed.
This means that when you press up and down with the keyboard to navigate the
accordion, whatever MS you hit will be the one displayed on the right, and
whatever one you were previously on will have its changes saved.

This is different from how DiskOverviews and SpokeSelectors work, but I think
it makes sense.  When you keyboard to a DO, you are choosing to do include it
in the install.  We don't want people doing that by accident.  When you
keyboard to a SS, it would jump you into the spoke and that wouldn't work
at all.  The MS, however, just changes what's displayed on part of the screen.