Use systemd to startup sshd when desired
This adds another chunk to the anaconda-generator that will figure out
what services go where.  In this case, we enable the sshd service when
inst.sshd or sshd is found as a command line option, or if we're on
s390x (because the console sucks there).
(cherry picked from commit fcb1049bffe200acebac3372d8af7fc60926c892)