Move sshpw handling out of Anaconda
This creates an ExecStartPre on the anaconda-sshd.service that will
check for a kickstart entry for sshpw.  It will add/modify users
accordingly prior to launching the sshd service.  Since now sshpw and
sshd bring up happens outside of and before anaconda starts, we can
remove sshd.py and any reference to it.