Don't start anaconda directly on s390x (#811477)
On s390x we want the user to ssh in to the install@ account.  This way
they have a functioning terminal in which to run the curses text install
or jump to the vnc install.  It also gives them a place where pdb will
work if they need to debug the install.

Some of the code here was grabbed from vnc.py to figure out the network
IP and name to prompt the user with.
(cherry picked from commit a4722359bfb154471ca0768452bea17d3493552d)