fix ks=cdrom (#828589)
We weren't handling the "empty $ksdev" case, which (as it turns out) is
legit for 'ks=cdrom'.

Note that the old option was "ks=cdrom[:<path>]". Rather than trying
to cleverly detect the difference between a device and path, we're
making it "cdrom:<dev>:<path>", which matches the "hd:.." argument.

Thus if you want to specify a path for the kickstart but not the CDROM
device, do: ks=cdrom::<path>

<path> defaults to '/ks.cfg', as always.
(cherry picked from commit 09d1ffd97dae66d5b8e6a3747de2976c88bfec76)