use timeout instead of idle for Queue polling
The Queue poll inside the GLib idle state was causing 100% CPU usage, it
was being called multiple times per second. It would be more efficient
to be able to use a select() or poll() with the Queue, but that isn't
possible. So, run the Queue processing in a 1 second timeout. CPU usage
drops to almost nothing.

Also add a call to q.task_done() for unrecognized messages.