Limit the maximum size of swap partitions
swapon creates a table using 1 B for each page in the swap. Therefore when a
huge swap (e.g. 500 GB) is created, swapon fails with ENOMEM, because it
is using vmalloc that allocates a continuous memory. Thus we need to limit the
maximum size of swap partitions. And since we don't know the strict limit, let's
use 128 GB that seems to work and should be enough for everybody.

(ported 35679a70eab52b3b040a61fce9d0e43dbb977c98 from rhel6-branch)