Allow quick deletion of an entire installation by clicking on a header.
This requires a little reorganization of the deletion code to make it reusable.
It also still goes through the confirmation dialog so the user has a choice to
think about what they're doing.

Note that it still doesn't cause a now-empty page to be removed (you can get
into that situation by removing everything yourself anyway), nor does it cause
anything useful to be displayed on the RHS, nor does it put back the
CreateNewPage if you delete the autopart-created installation.