Clean up disk (re)initialization code.
CLEARPART_TYPE_ALL means clear existing partitions and create a new
disklabel.

CLEARPART_TYPE_LINUX means clear linux-native partitions (or linux-
native whole-disk formatting) and create a new disklabel if the
removals result in an empty disk.

CLEARPART_TYPE_LIST means clear the specified devices (and devices
built on top of them) and create a new disklabel if the removals
result in an empty disk.

Zerombr becomes an alias for 'clearpart --initlabel'.

To initialize all disks:

  clearpart --all

To initialize uninitalized disks but otherwise clear no partitions:

  clearpart --none --initlabel

This adds a flag to the storage discovery config class that determines
whether or not clearPartitions will remove not-yet-created devices and
disklabels. The default is to not clear non-existent devices/formats.

After the first trip to the storage spoke, we should not remove any
layout the user set up in previous visits unless they choose autopart
again, in which case we start from scratch each time, or de-select
disks. Since the disk set can change each time through we have to be
able to initialize uninitialized disks even on the way to the custom ui.