Disk selection determines both ignoredisk.onlyuse and clearpart.drives.
This adds a way to hide devices from the devicetree without removing
them. The devicetree is initially populated without filtering because
we need to collect information about disks in the system as well as
free space in disks and filesystems. Disk selection in the storage
spoke establishes the set of disks to be used during install. We don't
want to re-populate the tree every time a disk is selected or
deselected. Instead, we just hide/unhide the disk and all devices it
contains.