Provide option to make non-zero exits fatal
execWithCapture normally doesn't care about non-zero exits in called
processes.  In some cases (most?) this is preferred.  In few cases we do
actually care about the success or failure of the execution.  "fatal"
provides a boolean that will allow callers to decide if they care or
not.

An OSError is raised because there is already code to catch that in the
except section.

Related: rhbz#730023
(cherry picked from commit ee29fa9f1861787fb8e173d22c3bd39bef15a26e)