Check passphrase unless explicit cancel (#821724)
snack/newt returns None for a button if F12 is pressed which is a global
default.  There can be passphrase data in the box when this happens and
we should read it.  The only time we should ignore anything in the box
is if the user has explicitely pressed the cancel button.  If the
passphrase box is empty but the user pressed F12 we still handle that
correctly as an implicit cancel.
(cherry picked from commit 682e5816cf38fdf641d8a9d89d2630009c07f67e)