Accept iSCSI details on F12 press (#823447)
When F12 (the global default continue hotkey) is pressed the button
value is None.  But we still want to process the details when this
happens.  The use of bool here makes it a little easier to read.
(cherry picked from commit c1fcbfa7593f8fdeffa5541dfa54e17d80f0c744)