Be honest about free space on disks.
We still do the final check against the software's disk space
requirement with the assumption that the user is doing autopart.

Even when we have the tug of war, we're still checking free space
before the user has had a chance to choose between autopart and
custom. That will need to be addressed at some point.