Make hub continue button sensitivity more complicated.
It can only be sensitive both if there's no incomplete spokes and no not
ready spokes.  By not tracking not ready spokes, it's possible to hit
continue while things are still downloading.  This can lead to very bad
situations, and it's only by accident that we've not hit this yet.