Fix auto-continuing for kickstart installs.
If you're on a spoke when all other spokes become ready and complete, the
install will continue automatically when you hit back.  If you're on the hub,
the install will continue automatically when all spokes are ready and
complete.

It's a little jarring, but that's just how it's gotta be.