remove Network(): list of network devices, final steps
This is almost the final step, the object is not used anymore,
I am keeping the rest only not to forget to handle what is left:
- I need to look at imageInstall
- creating default ifcfg files (setDefaultConfig) should go to
dracut or some network initialization step

The whole patchset removes anaconda.network object aiming to have all
data/configuration in ksdata.

The object was needed for our GUI using nm-c-e. We used to keep list of device
configuration objects (basically ifcfg dictionaries) in it so that we could
tweak them to be able to use nm-c-e for configuration/activation of devices.
Now it seems we can do without the list although we still need to do some
modifications of ifcfg files at the end of installation (setting onboot policy,
setting values for devices used for storage)

Hopefuly we'll be able to represent all the data stored in network object
in ksdata.

Following to the patchset I want to update ksdata with NetworkData objects
for all devices and use it to replace network.getDevices(). I have to think
where it should happen:
- in kickstart.py
- or somewhere in pre-gui network initialization (so that it happens also
  for non-ks cases) where we e.g. activate default device if needed.
- or in initialize of standalone spoke - seems to late, we'll need it
  already for eventual bring-up of network

-----

To sum up what the network.py serves for now:

Some of network utility functions (more of them are in isys):
- hostname sanity checking
- ip sanity checking
- hostname resolution
- status of networking
  - connected?
  - list of active devices
- logging (ifcfg files)

Network configuration:
- hostname setting (getting?)
  - note: storage (lvm, raid) is using hostname for default names
- ksdevice resolution (link, MAC address, bootif)
  - probably we'll be able to remove it, now it is only used
    for unspecified --device in kickstart network command
- write kickstart (currently from ifcfg config)
  - used by apply method
- write dracut arguments (from ifcfg config)
  - note: depends on storage
- modify configuration of target system (ifcfg files)
  - note: depends on storage
  - ONBOOT policy (differs on rhel and Fedora)
  - FCoE - ONBOOT=yes the devices
  - root on iSCSI - NM_CONTROLLED=no for root on iscsi
    (there is a NM BZ to fix this)
- write /etc/sysconfig/network configuration
  - this should be reviewed
- copy network configuration files to system
  - ifcfg-<iface> files and wireless key-<iface> files
  - dhclient-<iface>.conf files
    (dhcpclass and dhcp timeout which is not supported in noloader)
  - /etc/sysconfig/network
  - /etc/resolv.conf - genrated by NM
  - /etc/udev/rules.d/70-persistent-net.rules (review)
- disable ipv6 on target system (noipv6 boot/ks option)